kafka addition, and change the type of a & b to be more precise

+++ calculator-rest/src/test/java/Calculator/calculator/CalculatorControllerTest.java
@@ -30,9 +30,9 @@ public class CalculatorControllerTest {
     @Test
     public void testSum(){
         String operator = "sum";
-        when(calculatorService.calculating(operator, 3, 2)).thenReturn(5);
-        
-        int result = calculatorController.sum(3, 2);
+        when(calculatorService.calculating(operator, 3, 2)).thenReturn(5.0);
+
+        double result = calculatorController.sum(3, 2);
 
         assertEquals(5, result);
         verify(calculatorService, times(1)).calculating(operator, 3, 2);
@@ -41,9 +41,9 @@ public class CalculatorControllerTest {
     @Test
     public void testSub(){
         String operator = "sub";
-        when(calculatorService.calculating(operator, 5, 3)).thenReturn(2);
+        when(calculatorService.calculating(operator, 5, 3)).thenReturn(2.0);
 
-        int result = calculatorController.sub(5, 3);
+        double result = calculatorController.sub(5, 3);
         
         assertEquals(2, result);
         verify(calculatorService, times(1)).calculating(operator, 5, 3);
@@ -52,9 +52,9 @@ public class CalculatorControllerTest {
     @Test
     public void testDiv(){
         String operator = "div";
-        when(calculatorService.calculating(operator, 10, 2)).thenReturn(5);
+        when(calculatorService.calculating(operator, 10, 2)).thenReturn(5.0);
 
-        int result = calculatorController.div(10, 2);
+        double result = calculatorController.div(10, 2);
 
         assertEquals(5, result);
         verify(calculatorService, times(1)).calculating(operator, 10, 2);
@@ -63,9 +63,9 @@ public class CalculatorControllerTest {
     @Test
     public void testMult(){
         String operator = "mult";
-        when(calculatorService.calculating(operator, 10, 2)).thenReturn(20);
+        when(calculatorService.calculating(operator, 10, 2)).thenReturn(20.0);
 
-        int result = calculatorController.mult(10, 2);
+        double result = calculatorController.mult(10, 2);
 
         assertEquals(20, result);
         verify(calculatorService, times(1)).calculating(operator, 10, 2);
